refactored 404 page added loading page and started join page
